Create Mole Ngozichukwu Mary Nicolette-2024704207
This python program implements a grading system using 'if' and 'elif' statements. It takes a student's score as input and print the corresponding grade based on the following scale:
      -70-100: Grade A
      -60-69: Grade B
      -50-59: Grade C
       -40-49: Grade D
        -30-39: Grade E 
For each valid score, the program also prints "YES" to confirm the grading was successful. If the score is outside the valid range it prints "invalid score" and  "NO"

Created by: Mole Ngozichukwu Mary Nicolette
 Reg. No: 2024704207